requirements.txt
azure-functions
fastapi
python-dotenv
azure-identity
azure-ai-projects
azure-ai-agents
